requirements-dev.txt
bentoml
transformers
accelerate
optimum
pydantic
Pillow
open_clip_torch
torch
